studymatch/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.1.0.post1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis>=5.0.0
websockets==12.0
bcrypt==4.0.1
